% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{dd_expand_dotted}
\alias{dd_expand_dotted}
\title{Expand a flat dotted-key style file into a nested token list}
\usage{
dd_expand_dotted(x)
}
\description{
Expand a flat dotted-key style file into a nested token list
}
\keyword{internal}
