% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{designer_styles}
\alias{designer_styles}
\title{List available docdesigner styles}
\usage{
designer_styles()
}
\value{
A data frame of styles with the set they come from.
}
\description{
List available docdesigner styles
}
