% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{pdf}
\alias{pdf}
\title{Designed PDF output format}
\usage{
pdf(..., style = "minimal")
}
\arguments{
\item{...}{Passed to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}.}

\item{style}{A style from \code{\link[=designer_styles]{designer_styles()}} or a style directory path.}
}
\value{
An R Markdown output format.
}
\description{
Designed PDF output format
}
